// src/chains/functions/[Harmony-6] ViperSwap - Staking/data/interfaces.ts
export interface Contracts {
  Viper: string;
  MasterBreeder: string;
  ViperOneLP: string;
  Wswagmi: string;
  WswagmiOneLP: string;
  WswagmiViperLP: string;
  xViperTowsWAGMIAddress: string;
  wsWAGMIToViperAddress: string;
  xViperToViperAddress: string;
  xViperToCSHAREAddress: string;
  immortlOneToImmortlAddress: string;
  immortlViperToImmortlAddress: string;
  hplayOneToHPLAYAddress: string;
}

export interface OTFSettings {
  amtType: string;
  amt: number;
  tokenToDeposit: string;
  contractPreference: string;
  poolID: number;
}